+++ src/test/kotlin/org/jetbrains/changelog/tasks/PatchChangelogTaskTest.kt
@@ -113,9 +113,7 @@ class PatchChangelogTaskTest : BaseTest() {
         )
 
         assertFailsWith<MissingVersionException> {
-            extension.getUnreleased().also {
-                println("it = ${it}")
-            }
+            extension.getUnreleased()
         }
     }
 
@@ -604,6 +602,108 @@ class PatchChangelogTaskTest : BaseTest() {
     }
 
     @Test
+    fun `preserves complex list items`() {
+        changelog =
+            """
+            # Changelog
+            
+            ## [Unreleased]
+            
+            ### Changed
+            - This item has three paragraphs.
+            
+              All paragraphs should be preserved.
+            
+              Separate paragraphs must not be merged.
+            - This item contains a nested list.
+            
+              - Sub-Item 1
+              - Sub-Item 2
+            - ```
+              This item is a code block
+              ```
+            
+            ## [0.0.1] - 2022-10-10
+            
+            ### Added
+            - ```
+              A code block could also be followed by a list
+              ```
+            
+              - Sub-Item 1
+              - Sub-Item 2
+            
+              | Header                      |
+              |-----------------------------|
+              | There could also be a table |
+            
+            ### Fixed
+            - Following sections must stay unaffected.
+            """.trimIndent()
+
+        runTask(PATCH_CHANGELOG_TASK_NAME)
+
+        assertMarkdown(
+            """
+            # Changelog
+            
+            ## [Unreleased]
+            
+            ### Added
+            
+            ### Changed
+            
+            ### Deprecated
+            
+            ### Removed
+            
+            ### Fixed
+            
+            ### Security
+            
+            ## [1.0.0] - $date
+            
+            ### Changed
+            - This item has three paragraphs.
+            
+              All paragraphs should be preserved.
+            
+              Separate paragraphs must not be merged.
+            - This item contains a nested list.
+            
+              - Sub-Item 1
+              - Sub-Item 2
+            - ```
+              This item is a code block
+              ```
+            
+            ## [0.0.1] - 2022-10-10
+            
+            ### Added
+            - ```
+              A code block could also be followed by a list
+              ```
+            
+              - Sub-Item 1
+              - Sub-Item 2
+            
+              | Header                      |
+              |-----------------------------|
+              | There could also be a table |
+            
+            ### Fixed
+            - Following sections must stay unaffected.
+            
+            [Unreleased]: https://github.com/JetBrains/gradle-changelog-plugin/compare/v1.0.0...HEAD
+            [1.0.0]: https://github.com/JetBrains/gradle-changelog-plugin/compare/v0.0.1...v1.0.0
+            [0.0.1]: https://github.com/JetBrains/gradle-changelog-plugin/commits/v0.0.1
+            
+            """.trimIndent(),
+            changelog
+        )
+    }
+
+    @Test
     fun `removes empty groups`() {
         changelog =
             """
@@ -1013,6 +1113,65 @@ class PatchChangelogTaskTest : BaseTest() {
             """.trimIndent(),
             extension.render()
         )
+    }
+
+
+    @Test
+    fun `do not break configuration cache with custom sectionUrlBuilder`() {
+        buildFile =
+            """
+            import org.jetbrains.changelog.ChangelogSectionUrlBuilder
+            
+            plugins {
+                id 'org.jetbrains.changelog'
+            }
+            changelog {
+                version = "1.0.0"
+                repositoryUrl = "https://github.com/JetBrains/gradle-changelog-plugin"
+                sectionUrlBuilder = { repositoryUrl, currentVersion, previousVersion, isUnreleased ->
+                    "repositoryUrl=" + repositoryUrl + "|currentVersion=" + currentVersion + "|previousVersion=" + previousVersion + "|isUnreleased=" + isUnreleased
+                } as ChangelogSectionUrlBuilder
+            }
+            """.trimIndent()
+
+        project.evaluate()
+        runTask(PATCH_CHANGELOG_TASK_NAME)
+        val result = runTask(PATCH_CHANGELOG_TASK_NAME)
+
+        assertMarkdown(
+            """
+            <!-- Foo bar -->
+            
+            # Changelog
+            My project changelog.
+            
+            ## [Unreleased]
+            
+            ### Added
+            
+            ### Changed
+            
+            ### Deprecated
+            
+            ### Removed
+            
+            ### Fixed
+            
+            ### Security
+            
+            ## [1.0.0] - $date
+            Fancy release.
+            
+            ### Added
+            - foo
+            
+            [Unreleased]: repositoryUrl=https://github.com/JetBrains/gradle-changelog-plugin|currentVersion=Unreleased|previousVersion=1.0.0|isUnreleased=true
+            [1.0.0]: repositoryUrl=https://github.com/JetBrains/gradle-changelog-plugin|currentVersion=1.0.0|previousVersion=null|isUnreleased=false
+            """.trimIndent(),
+            extension.render()
+        )
+
+        assertTrue(result.output.contains("Reusing configuration cache."))
     }
 
     @Test