+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Added
+- Add `--no-empty-sections` flag to `getChangelog` task. [#167](../../issues/167)
+
 ### Fixed
 - No-longer discard all but the last paragraph in list items [#133](../../issues/133) [#147](../../issues/147)