+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 - Add `--no-empty-sections` flag to `getChangelog` task. [#167](../../issues/167)
 
+### Fixed
+- No-longer discard all but the last paragraph in list items [#133](../../issues/133) [#147](../../issues/147)
+
 ## [2.0.0] - 2022-10-28
 
 ### Added